zadania/Z0201/requirements.txt
requests
pandas
python-dotenv
tiktoken
mcp